added sweet alert for doctors delete confirmation

+++ src/Pages/Appointment/Dashboard/DoctorsRow.js
@@ -1,34 +1,50 @@
 import React from 'react';
 import { toast } from 'react-toastify';
+import swal from 'sweetalert';
 
-const DoctorsRow = ({ doctor, index , refetch}) => {
-  const { name, specialization, img , email} = doctor;
+const DoctorsRow = ({ doctor, index, refetch }) => {
+  const { name, specialization, img, email } = doctor;
 
   const handleDelete = (email) => {
-    fetch(`http://localhost:5000/doctor/${email}`, {
-      method: 'DELETE',
-      headers: {
-        authorization: `Bearer ${localStorage.getItem('accessToken')}`
+    swal({
+      title: 'Are you sure?',
+      text: `You want to delete ${name}? `,
+      icon: 'warning',
+      buttons: true,
+      dangerMode: true,
+    }).then((willDelete) => {
+      if (willDelete) {
+        swal(`Doctor: ${name} is deleted`, {
+          icon: 'success',
+        });
+
+        fetch(`http://localhost:5000/doctor/${email}`, {
+          method: 'DELETE',
+          headers: {
+            authorization: `Bearer ${localStorage.getItem('accessToken')}`,
+          },
+        })
+          .then((res) => res.json())
+          .then((data) => {
+            console.log(data);
+            if (data.deletedCount) {
+              // toast.success(`Doctor: ${name} is deleted`);
+              refetch();
+            }
+          });
+      } else {
+        // swal('Your Doctors is safe!');
       }
-    })
-    .then(res => res.json())
-    .then(data => {
-      console.log(data);
-      if(data.deletedCount){
-        toast.success(`Doctor: ${name} is deleted`)
-        refetch();
-      }
-    })
-  }
-  
-  
+    });
+  };
+
   return (
     <tr>
       <th>{index + 1}</th>
       <td>
         <div class="avatar">
           <div class="w-16 rounded-full ring ring-primary ring-offset-base-100 ring-offset-2">
-            <img src={img} alt=''/>
+            <img src={img} alt="" />
           </div>
         </div>
       </td>
@@ -36,7 +52,12 @@ const DoctorsRow = ({ doctor, index , refetch}) => {
       <td>{specialization}</td>
       <td>
         {' '}
-        <button onClick={()=> handleDelete(email)} className="btn btn-xs btn-error">Delete</button>{' '}
+        <button
+          onClick={() => handleDelete(email)}
+          className="btn btn-xs btn-error"
+        >
+          Delete
+        </button>{' '}
       </td>
     </tr>
   );
